Reorganize and reword the documentation
The command line tool is the primary usage, so try to promote
information as first thing that is emphasised, while retaining
all the other information for people who want to get deeper.

Move the TypeTree information to a dedication page (which
may expand a lot over time)

Also a few corrections.