WIP Attempt to support analysis of more than one build
If two versions of the same builds are in separate directories then we can potentially support analysis into a single DB by incorporating the path into the unique id generation.
However so far this approach is breaking the asset dependency view, so either this approach is fundamentally flawed or a few more fixes could complete it.

Even with this fix dependencies between AssetBundles are fundamentally not "1-1" if different AssetBundles have the same filenames and are in the same folder - we don't know which one to pick.  This can be the case with AssetBundle variants or other cases where the game logic konws which AssetBundle to load at runtime and only one of the "conflicting" bundles will be loaded.  How analyze should properly handle this case is TBD.