Analyze command - Improvements in error handling and logging (#34)
* Modest improvement in error reporting and Archive handling

Check for Archives rather than trying to mount every file just in case it turns out to be an archive.
Unity Archives have a distinctive signature string at the start so we can pre-triage which files are actually Unity Archives.
The expected values were confirmed by studying the Unity parsing code and debugging the load of some files in the editor.

Output improvement - when erasing progress message move back to the start of the line so that errors are not printed at a random indent.

This is only a modest improvement because we still don't have some mechanism to predict if a file is a Unity Serialized File without attempting to load it and then failing.

* Analyze - summarize number of successful, ignored and failed files

Previously it was hard to tell if errors were shown whether anything actually had been analyzed.
Partial imports are normal until we find a way to reliably detect serialized files.

* Document more expected errors from Analyze

Mention errors that migth be visible (based on testing a folder with AssetBundles from different builds all being mixed together)