Establish version number
Set current version to 1.1.0

To see the version run
UnityDataTool --version

Example output:
1.1.0+88c405f51c46f2191305bbd204748a68d7bb94cf

Prior to this commit, "--version" would print "1.0.0" (the default value)
So, although there has never been an official "1.0.0" version, we will start to increment through 1.1, 1.2, 1.3 going forward.

The git commit is automatically included in the version which will help distinguish if we neglect to upgrade the version when adding improvements.