Restore "assets" population and documentation updates
The AssetBundleHandler got lost in recent code change so restore it.

Add a few cross links to the new Addressables Build Layout parsing.

Fix the section "Matching content back to the source asset" because it wasn't explaining that some source asset path are available in the build output (for AssetBundle explicit assets, and Player build scenes)