Restore "assets" table population and documentation tweaks (#38)
The AssetBundleHandler got lost in recent code change, so restore it.

Add a few cross links to the new Addressables Build Layout parsing.

Fix the section "Matching content back to the source asset" because it wasn't explaining that some source asset paths actually are available in the build output (for AssetBundle explicit assets, and the scenes in a Player Build)

New example showing how to find implicitly referenced assets.