And initial test for Addressables BuildLayout import support
Using two buildlayout files from building the AudioExample project.

The test does some checks of hardcoded values across some of the main tables, to confirm that the database tables were populated as expected.
This is just a "sanity check", but demonstrates how anything can be checked in the the reference data.